Overhaul, time! (vagrant, ruby, etc). Fixes #30
I wanted to redirect /resume to my StackOverflow Careers page, but the Jekyll plugin required a newer version of the Jekyll gem. I had to update large chunks of the Gemfile. Then, I noticed the local Vagrant machine was too complicated, so I updated that to use distribution packages only.